Convert LeftColumn type from `sbyte` to `byte` -- this column must be present !

+++ GeUtilities.Tests/BEDParserTests/TestColumnOrders.cs
@@ -38,7 +38,7 @@ namespace GeUtilities.Tests.BEDParserTests
         [InlineData(4, 3, 2, 1, 0)]
         [InlineData(2, 1, 0, 4, 3)]
         [InlineData(5, 6, 7, 8, 9)]
-        public void TestColumnsShuffle(sbyte chrColumn, sbyte leftColumn, sbyte rightColumn, byte nameColumn, byte valueColumn)
+        public void TestColumnsShuffle(sbyte chrColumn, byte leftColumn, sbyte rightColumn, byte nameColumn, byte valueColumn)
         {
             int left = 10, right = 20;
             string name = "GeUtilities_00", peak = "";
@@ -110,7 +110,7 @@ namespace GeUtilities.Tests.BEDParserTests
                 BEDParser<ChIPSeqPeak> bedParser = new BEDParser<ChIPSeqPeak>(
                     testFile.TempFilePath,
                     chrColumn: chrColumn,
-                    leftEndColumn: leftColumn,
+                    leftEndColumn: (byte)leftColumn,
                     rightEndColumn: rightColumn,
                     nameColumn: (byte)nameColumn,
                     valueColumn: (byte)valueColumn,
@@ -161,7 +161,7 @@ namespace GeUtilities.Tests.BEDParserTests
                 BEDParser<ChIPSeqPeak> bedParser = new BEDParser<ChIPSeqPeak>(
                     testFile.TempFilePath,
                     chrColumn: chrColumn,
-                    leftEndColumn: leftColumn,
+                    leftEndColumn: (byte)leftColumn,
                     rightEndColumn: rightColumn,
                     nameColumn: (byte)nameColumn,
                     valueColumn: (byte)valueColumn,
